Initial work towards Win64 compilation.
This is nowhere near building, let alone working, but we'll get there
eventually.

+++ neo/idlib/Heap.cpp
@@ -42,7 +42,7 @@ If you have questions concerning this license or the applicable additional terms
 Mem_Alloc16
 ==================
 */
-void * Mem_Alloc16( const int size, const memTag_t tag ) {
+void * Mem_Alloc16( const size_t size, const memTag_t tag ) {
 	if ( !size ) {
 		return NULL;
 	}
@@ -67,7 +67,7 @@ void Mem_Free16( void *ptr ) {
 Mem_ClearedAlloc
 ==================
 */
-void * Mem_ClearedAlloc( const int size, const memTag_t tag ) {
+void * Mem_ClearedAlloc( const size_t size, const memTag_t tag ) {
 	void * mem = Mem_Alloc( size, tag );
 	SIMDProcessor->Memset( mem, 0, size );
 	return mem;
